split `unnecessary_clone` test into `clone_on_ref_ptr` and `clone_on_copy`
